requirements.txt
# N8 Gateway - Python Dependencies
# FastAPI核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
# 数据库
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy[asyncio]==2.0.23
# 认证
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pydantic==2.5.2
pydantic-settings==2.1.0
# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
# 工具
python-dotenv==1.0.0
httpx==0.25.2
